+++ src/collective/contact/plonegroup/interfaces.py
@@ -23,5 +23,30 @@ class INotPloneGroupContact(Interface):
 
 
 class IPlonegroupGroupCreatedEvent(IPASEvent):
-    """A new Plone group linked to an organization has been created.
+    """
+        A new Plone group linked to an organization has been created.
+    """
+
+
+class IGroupField(Interface):
+    """
+        Interface for the GroupField
+    """
+
+
+class IOrganizationField(Interface):
+    """
+        Interface for the OrganizationField
+    """
+
+
+class IDGFVocabularyField(Interface):
+    """
+        Interface for the DGFVocabularyField
+    """
+
+
+class IDGFListField(Interface):
+    """
+        Interface for the DGFListField
     """