+++ CHANGES.rst
@@ -1,12 +1,16 @@
 Changelog
 =========
 
-
 1.30 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added manage-own-groups-users view and functions/groups selection configuration.
+  A user can manage the user assignments of his groups.
+  [sgeulette]
+- Avoid getting groups (only users) in `get_selected_org_suffix_users` function.
+  [sgeulette]
+- Avoid GroupNotFoundError in `get_selected_org_suffix_users` if suffix is limited to some organizations.
+  [sgeulette]
 
 1.29.1 (2020-08-26)
 -------------------